refactored routes and added controllers, updated client API calls,handeled user register error alert
