+++ routes/authRoutes.js
@@ -1,0 +1,50 @@
+const passport = require("../config/passport");
+const router = require("express").Router();
+const auth = require("../controllers/auth-controller");
+
+
+router.route("/band")
+  .post( function (req, res, next) {
+    console.log("routes/user.js, login, req.body: ");
+    console.log(req.body);
+    next();
+  },
+  passport.authenticate("band-local"),
+  (req, res) => {
+    console.log("logged in", req.user.userName);
+    var userInfo = {
+      username: req.user.userName,
+      sessionId: req.sessionID,
+      id: req.user.id
+    };
+    res.send(userInfo);
+  },
+  )
+  .put()
+  .delete(auth.deleteBand);
+
+router.route("/user")
+  .post( function (req, res, next) {
+    console.log("routes/user.js, login, req.body: ");
+    console.log(req.body);
+    next();
+  },
+  passport.authenticate("user-local"),
+  (req, res) => {
+    console.log("logged in",req.user.userName);
+    var userInfo = {
+      username: req.user.userName,
+      sessionId: req.sessionID,
+      id: req.user.id
+    };
+    res.send(userInfo);
+  }
+  )
+  .put()
+  .delete(auth.deleteUser);
+
+router.route("/logout")
+  .get(auth.logout);
+
+
+module.exports = router;
