+++ server.js
@@ -8,9 +8,10 @@ const PORT = process.env.PORT || 3001;
 const app = express();
 const db = require("./models");
 // require routes
-const bandAuthRoutes = require("./routes/bandAuthRoutes");
-const bandApiRoutes = require("./routes/bandInfoRoutes");
-const userAuthRoutes = require("./routes/userAuthRoutes");
+const authRoutes = require("./routes/authRoutes");
+const userApiRoutes = require("./routes/userApiRoutes");
+const bandApiRoutes = require("./routes/bandApiRoutes");
+
 // Define middleware here
 app.use(compression());
 app.use(express.urlencoded({ extended: true }));
@@ -30,10 +31,9 @@ app.use(passport.session());
 
 
 // Use apiRoutes
+app.use("/api", userApiRoutes,bandApiRoutes);
+app.use("/auth", authRoutes);
 
-app.use("/auth/band", bandAuthRoutes);
-app.use("/api", bandApiRoutes);
-app.use("/auth", userAuthRoutes);
 // Send every request to the React app
 app.get("*", function (req, res) {
   res.sendFile(path.join(__dirname, "./client/build/index.html"));
